meta settingd completed
